directories redefined
